# modules/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: BE Admin Workspace
  version: '1.0'
servers:
  - url: http://localhost:6303
    description: Local Server
paths:
...
  /organizations:
    get:
      summary: Get organizations
      description: Retrieve a list of organizations
      parameters:
        - name: limit
          in: query
          description: Number of organizations to retrieve (default is 30)
          required: false
          type: integer
        - name: offset
          in: query
          description: Number of organizations to skip (default is 0)
          required: false
          type: integer
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              count:
                type: integer
              results:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '500':
          description: Internal server error

    post:
      summary: Create organization
      description: Create a new organization
      parameters:
        - name: organizationData
          in: body
          description: Organization data
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Organization created successfully
          schema:
            $ref: '#/components/schemas/Organization'
        '400':
          description: Bad request
        '500':
          description: Internal server error
...
components:
  schemas:
...
    Organization:
      type: object
      properties:
        organization_id:
          type: integer
          format: int64
          example: 1
        organization_name:
          type: string
          maxLength: 255
          example: Example Organization
        status:
          type: string
          enum: ['active', 'inactive'] # Assuming organization_status_type has these values
          example: active
        address:
          type: string
          maxLength: 255
          example: 123 Main St
        contact_number:
          type: string
          maxLength: 255
          example: '+1234567890'
        email:
          type: string
          format: email
          maxLength: 255
          example: info@example.com
        country:
          type: string
          maxLength: 255
          example: United States
        state:
          type: string
          maxLength: 255
          example: California
        city:
          type: string
          maxLength: 255
          example: San Francisco
        postal_code:
          type: string
          maxLength: 255
          example: '12345'
        website_url:
          type: string
          format: uri
          maxLength: 255
          example: https://www.example.com
        kvp:
          type: object
          example: { 'key': 'value' } # Example JSONB structure
        date_created:
          type: string
          format: date-time
          example: '2023-10-17T12:00:00Z'
        date_modified:
          type: string
          format: date-time
          example: '2023-10-17T12:30:00Z'
      required:
        - organization_name
        - status
    OrganizationCreate:
      type: object
      properties:
        organization_name:
          type: string
          maxLength: 255
          example: Example Organization
        status:
          type: string
          enum: ['active', 'inactive']
          example: active
        address:
          type: string
          maxLength: 255
          example: 123 Main St
        contact_number:
          type: string
          maxLength: 255
          example: '+1234567890'
        email:
          type: string
          format: email
          maxLength: 255
          example: info@example.com
        country:
          type: string
          maxLength: 255
          example: United States
        state:
          type: string
          maxLength: 255
          example: California
        city:
          type: string
          maxLength: 255
          example: San Francisco
        postal_code:
          type: string
          maxLength: 255
          example: '12345'
        website_url:
          type: string
          format: uri
          maxLength: 255
          example: https://www.example.com
        kvp:
          type: object
          example: { 'key': 'value' } # Example JSONB structure
      required:
        - organization_name
        - status
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []